+++ xskillscore/tests/test_effective_p_value.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pandas as pd
 import pytest
 import xarray as xr
 
@@ -16,7 +15,7 @@ DIM = "time"
 
 @pytest.fixture
 def a():
-    times = pd.date_range("1/1/2000", "2/3/2000", freq="D")
+    times = xr.cftime_range("2000-01-01", "2000-01-03", freq="D")
     lats = np.arange(4)
     lons = np.arange(5)
     data = np.random.rand(len(times), len(lats), len(lons))