+++ xskillscore/tests/test_metric_results_accurate.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pandas as pd
 import pytest
 from scipy.stats import pearsonr, spearmanr
 from sklearn.metrics import (
@@ -28,7 +27,7 @@ from xskillscore.core.deterministic import (
 
 @pytest.fixture
 def a():
-    time = pd.date_range("1/1/2000", "1/3/2000", freq="D")
+    time = xr.cftime_range("2000-01-01", "2000-01-03", freq="D")
     da = xr.DataArray(np.random.rand(len(time)), dims=["time"], coords=[time])
     return da